+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,15 +2,15 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_194467: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_194487: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_192342: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_192353: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_199168: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_198965: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_199166: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_199178: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_199164: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_199417: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_199216: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_194718: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_194738: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199419: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199429: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_192593: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_192604: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199415: (a: number, b: number, c: number, d: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;