+++ src/game/units/archer/plugin.rs
@@ -1,7 +1,8 @@
 use bevy::prelude::*;
 
+use super::components::{Archer, Arrow};
 use super::systems::*;
-use crate::game::run_conditions;
+use crate::game::run_conditions::{any_exist, coming_from_game_over};
 use crate::state::{AppState, InGameState};
 
 pub struct ArcherPlugin;
@@ -21,26 +22,28 @@ impl Plugin for ArcherPlugin {
                 spawn_initial_defender_archers,
                 spawn_initial_attacker_archers,
             )
-                .run_if(run_conditions::coming_from_game_over),
-        )
-        .add_systems(
-            Update,
-            update_archer_targeting.in_set(crate::game::plugin::VelocitySystemSet),
-        )
-        .add_systems(
-            Update,
-            archer_movement.in_set(crate::game::units::MovementCalculationSet),
+                .run_if(coming_from_game_over),
         )
         .add_systems(
             Update,
             (
-                update_archer_movement_timers,
-                archer_melee_combat,
-                archer_ranged_combat,
-                move_arrows,
-                check_arrow_collisions,
+                update_archer_targeting.in_set(crate::game::plugin::VelocitySystemSet),
+                archer_movement.in_set(crate::game::units::MovementCalculationSet),
+                (
+                    update_archer_movement_timers,
+                    archer_melee_combat,
+                    archer_ranged_combat,
+                )
+                    .chain(),
             )
+                .run_if(any_exist::<Archer>())
+                .run_if(in_state(InGameState::Running)),
+        )
+        .add_systems(
+            Update,
+            (move_arrows, check_arrow_collisions)
                 .chain()
+                .run_if(any_exist::<Arrow>())
                 .run_if(in_state(InGameState::Running)),
         );
     }