+++ src/game/units/king/plugin.rs
@@ -1,11 +1,11 @@
 use bevy::prelude::*;
 
 use crate::game::plugin::{MovementSystemSet, VelocitySystemSet};
-use crate::game::run_conditions;
+use crate::game::run_conditions::{any_exist, coming_from_game_over};
 use crate::game::shared_systems::apply_separation;
 use crate::state::{AppState, InGameState};
 
-use super::components::KingSpawned;
+use super::components::{King, KingSpawned};
 use super::systems;
 
 pub struct KingPlugin;
@@ -16,28 +16,22 @@ impl Plugin for KingPlugin {
             .add_systems(OnEnter(AppState::InGame), systems::spawn_king)
             .add_systems(
                 OnEnter(InGameState::Running),
-                systems::spawn_king.run_if(run_conditions::coming_from_game_over),
+                systems::spawn_king.run_if(coming_from_game_over),
             )
             .add_systems(
                 Update,
-                systems::update_king_targeting.in_set(VelocitySystemSet),
-            )
-            .add_systems(
-                Update,
-                systems::king_movement.in_set(crate::game::units::MovementCalculationSet),
-            )
-            .add_systems(
-                Update,
-                systems::king_cohesion_aura
-                    .after(apply_separation)
-                    .before(MovementSystemSet)
+                (
+                    systems::update_king_targeting.in_set(VelocitySystemSet),
+                    systems::king_movement.in_set(crate::game::units::MovementCalculationSet),
+                    systems::king_cohesion_aura
+                        .after(apply_separation)
+                        .before(MovementSystemSet),
+                    systems::snap_kings_guard_to_king
+                        .in_set(MovementSystemSet)
+                        .after(systems::king_movement),
+                )
+                    .run_if(any_exist::<King>())
                     .run_if(in_state(InGameState::Running)),
-            )
-            .add_systems(
-                Update,
-                systems::snap_kings_guard_to_king
-                    .in_set(MovementSystemSet)
-                    .after(systems::king_movement),
             );
     }
 }