+++ src/game/units/wizard/spells/finger_of_death/plugin.rs
@@ -2,6 +2,7 @@ use bevy::prelude::*;
 
 use super::super::super::components::Spell;
 use super::super::run_conditions::*;
+use super::components::FingerOfDeathBeam;
 use super::systems::*;
 use crate::state::InGameState;
 
@@ -17,11 +18,14 @@ impl Plugin for FingerOfDeathPlugin {
                     .run_if(spell_input_not_blocked)
                     .run_if(mouse_left_not_consumed)
                     .run_if(mouse_held_or_wizard_casting),
-                apply_finger_of_death_damage,
-                update_finger_of_death_beam_visuals,
-                cleanup_finger_of_death_beams,
+                (
+                    apply_finger_of_death_damage,
+                    update_finger_of_death_beam_visuals,
+                    cleanup_finger_of_death_beams,
+                )
+                    .chain()
+                    .run_if(any_exist::<FingerOfDeathBeam>()),
             )
-                .chain()
                 .run_if(in_state(InGameState::Running)),
         );
     }