+++ src/game/units/wizard/spells/guardian_circle/plugin.rs
@@ -2,6 +2,7 @@ use bevy::prelude::*;
 
 use super::super::super::components::Spell;
 use super::super::run_conditions::*;
+use super::components::GuardianCircleIndicator;
 use super::systems;
 use crate::state::InGameState;
 
@@ -24,9 +25,8 @@ impl Plugin for GuardianCirclePlugin {
                     .run_if(spell_input_not_blocked)
                     .run_if(mouse_left_not_consumed)
                     .run_if(mouse_held_or_wizard_casting),
-                systems::update_circle_indicator,
+                systems::update_circle_indicator.run_if(any_exist::<GuardianCircleIndicator>()),
             )
-                .chain()
                 .run_if(in_state(InGameState::Running)),
         );
     }