+++ src/game/units/wizard/spells/teleport/plugin.rs
@@ -4,6 +4,7 @@ use bevy::prelude::*;
 
 use super::super::super::components::Spell;
 use super::super::run_conditions::*;
+use super::components::{TeleportDestinationCircle, TeleportSourceCircle};
 use super::systems;
 use crate::state::InGameState;
 
@@ -27,7 +28,10 @@ impl Plugin for TeleportPlugin {
                     .run_if(mouse_left_not_consumed)
                     .run_if(mouse_right_not_held)
                     .run_if(mouse_held_or_wizard_casting),
-                systems::update_circle_animations,
+                systems::update_circle_animations.run_if(
+                    any_exist::<TeleportDestinationCircle>()
+                        .or(any_exist::<TeleportSourceCircle>()),
+                ),
             )
                 .run_if(in_state(InGameState::Running)),
         );